+++ src/operation/iRCX/source/data_manager/advance/LayerShape.hpp
@@ -24,19 +24,19 @@ class LayerShape
 {
  public:
   LayerShape() = default;
-  LayerShape(size_t layer_id, const GTLRectInt& shape) : _layer_id(layer_id), _shape(shape) {}
+  LayerShape(int32_t layer_id, const GTLRectInt& shape) : _layer_id(layer_id), _shape(shape) {}
   ~LayerShape() = default;
   // getter
-  size_t get_layer_id() const { return _layer_id; }
+  int32_t get_layer_id() const { return _layer_id; }
   GTLRectInt& get_shape() { return _shape; }
   const GTLRectInt& get_shape() const { return _shape; }
   // setter
-  void set_layer_id(size_t layer_id) { _layer_id = layer_id; }
+  void set_layer_id(int32_t layer_id) { _layer_id = layer_id; }
   void set_shape(const GTLRectInt& shape) { _shape = shape; }
   // function
 
  private:
-  size_t _layer_id = SIZE_MAX;
+  int32_t _layer_id = INT32_MAX;
   GTLRectInt _shape = GTLRectInt();
 };
