+++ src/operation/iRCX/source/module/topo_builder/TopoBuilder.hpp
@@ -52,10 +52,11 @@ class TopoBuilder
   void buildAll();
   TBNetTopo buildNet(Net& net);
   void appendNodeIfAbsent(Net& net, std::vector<TopoNode>& node_list, std::vector<bool>& node_shape_valid_list,
-                          std::map<std::tuple<size_t, int32_t, int32_t>, size_t>& node_key_to_idx_map,
-                          std::map<std::string, bool>& pin_consumed_map, size_t layer_id, const GTLPointInt& point);
-  size_t appendNode(std::vector<TopoNode>& node_list, std::vector<bool>& node_shape_valid_list, TopoNode node, bool is_shape_valid);
-  void mergeNodeShape(std::vector<TopoNode>& node_list, std::vector<bool>& node_shape_valid_list, size_t node_idx, const GTLRectInt& shape);
+                          std::map<std::tuple<int32_t, int32_t, int32_t>, int32_t>& node_key_to_idx_map,
+                          std::map<std::string, bool>& pin_consumed_map, int32_t layer_id, const GTLPointInt& point);
+  int32_t appendNode(std::vector<TopoNode>& node_list, std::vector<bool>& node_shape_valid_list, TopoNode node, bool is_shape_valid);
+  void mergeNodeShape(std::vector<TopoNode>& node_list, std::vector<bool>& node_shape_valid_list, int32_t node_idx,
+                      const GTLRectInt& shape);
   GTLRectInt getEndpointShape(Segment& segment, const GTLPointInt& point);
   void buildSpecial();
 };
